+++ src/include/common/string.h
@@ -28,7 +28,7 @@ extern bool pg_str_endswith(const char *str, const char *end);
 extern int	strtoint(const char *pg_restrict str, char **pg_restrict endptr,
 					 int base);
 extern char *pg_clean_ascii(const char *str, int alloc_flags);
-extern int	pg_strip_crlf(char *str);
+extern size_t pg_strip_crlf(char *str);
 extern bool pg_is_ascii(const char *str);
 
 /* functions in src/common/pg_get_line.c */
